server/caserver/requirements.txt
APScheduler==3.6.3
asgiref==3.2.7
beautifulsoup4==4.8.2
bs4==0.0.1
certifi==2019.11.28
chardet==3.0.4
Django==3.0.4
django-apscheduler==0.3.0
django-cors-headers==3.2.1
django-cors-middleware==1.5.0
django-filter==2.2.0
djangorestframework==3.11.0
idna==2.8
Markdown==3.1
pytz==2019.3
requests==2.22.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
tzlocal==2.0.0
urllib3==1.25.8
